requirements.txt
transformers>=4.30.0
torch>=2.0.0
markdown
logging
reportlab
fpdf>=1.7.2
accelerate>=0.3.0
python-dotenv